refactor import paths for PhysicsParameters to align with new directory structure
