+++ src/pleiades/sammy/fitting/config.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from pydantic import BaseModel, Field, constr
 from pleiades.nuclear.parameters import nuclearParameters
-from pleiades.core.experimental import PhysicsParameters
+from pleiades.experimental.parameters import PhysicsParameters
 from pleiades.sammy.data.options import dataParameters
 from pleiades.sammy.fitting.options import FitOptions
